First instance of state memory

+++ Net2Plan-GUI/Net2Plan-GUI-Plugins/Net2Plan-NetworkDesign/src/main/java/com/net2plan/gui/plugins/networkDesign/topologyPane/jung/osmSupport/state/CanvasStateMirror.java
@@ -11,9 +11,9 @@ class CanvasStateMirror
     private final CanvasState state;
 
     private final Point2D canvasCenter;
-    private final int zoomLevel;
+    private final double zoomLevel;
 
-    CanvasStateMirror(CanvasState state, Point2D mapCenter, int zoomLevel)
+    CanvasStateMirror(CanvasState state, Point2D mapCenter, double zoomLevel)
     {
         this.state = state;
         this.canvasCenter = mapCenter;
@@ -30,7 +30,7 @@ class CanvasStateMirror
         return canvasCenter;
     }
 
-    int getZoomLevel()
+    double getZoomLevel()
     {
         return zoomLevel;
     }
